requirements.txt
future==0.18.2
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.0
torch==1.6.0
torchvision==0.7.0
